Blog Website Recreation Project
This project is part of a roadmap I'm following to learn JavaScript. The challenge was to recreate a blog website from scratch, which has been instrumental in building foundational skills for my journey in web development. Through this project, I focused on styling the website with CSS and implemented functionality to dynamically populate the "blog content" using JavaScript. This hands-on experience has significantly contributed to my understanding of JavaScript and its practical applications in web development.